Yuna-Ishikawa/requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.6.3
apt-xapian-index==0.49
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.4
beautifulsoup4==4.11.1
BingImageCreator==0.5.0
blessed==1.20.0
blinker==1.6.3
Brlapi==0.8.4
Brotli==1.0.9
browser-cookie3==0.19.1
bs4==0.0.1
cachetools==4.2.2
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
chromedriver-autoinstaller==0.4.0
clean-text==0.6.0
cleantext==1.1.4
cleverdict==1.9.2
click==8.1.7
colorama==0.4.6
command-not-found==0.3
cryptography==38.0.4
cupshelpers==1.0
curl-cffi==0.5.9
dbus-python==1.3.2
decouple==0.0.7
defer==1.0.6
distro==1.8.0
distro-info==1.5
dnspython==2.3.0
duckduckgo-search==3.9.3
EdgeGPT==0.13.2
emoji==1.7.0
exceptiongroup==1.0.1
fake-useragent==0.1.14
fastapi==0.104.1
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
ftfy==6.1.1
fuse-python==1.0.5
g4f==0.1.8.1
google-ai-generativelanguage==0.3.3
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-generativeai==0.2.1
googleapis-common-protos==1.61.0
GoogleBard==2.1.0
gpg==1.18.0
grpcio==1.59.0
grpcio-status==1.59.0
gspread==5.12.0
gTTS==2.4.0
gyp==0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httplib2==0.20.4
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==4.12.0
iniconfig==2.0.0
inquirer==3.1.3
itsdangerous==2.1.2
jaraco.classes==3.2.1
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.3.2
Js2Py==0.74
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kaitaistruct==0.10
keyboard==0.13.5
keyring==23.9.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
linkedin-scraper==2.9.1
log2d==0.0.18
loguru==0.7.2
louis==3.24.0
lxml==4.9.3
lz4==4.3.2
Mako==1.2.4.dev0
Markdown==3.4.3
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mdurl==0.1.2
meson==1.0.1
more-itertools==8.10.0
MouseInfo==0.1.3
multidict==6.0.4
multipledispatch==0.6.0
nest-asyncio==1.5.8
netifaces==0.11.0
nltk==3.8.1
notify2==0.3
nqntnqnqmb==1.1
numpy==1.23.4
oauth2client==4.1.3
oauthlib==3.2.2
olefile==0.46
openai==0.28.1
OpenAIAuth==3.0.0
outcome==1.2.0
packaging==23.2
pandas==1.5.1
pexpect==4.8.0
Pillow==9.4.0
platformdirs==3.11.0
pluggy==1.3.0
progress==1.6
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.53
pycairo==1.20.1
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.19.0
pycups==2.0.1
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
PyExecJS==1.5.1
PyGetWindow==0.0.9
Pygments==2.14.0
PyGObject==3.44.1
pyjsparser==2.7.1
PyJWT==2.6.0
pylibacl==0.6.0
PyMsgBox==1.0.9
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.9
PyQt5-sip==12.11.1
PyRect==0.2.0
pyrsistent==0.19.3
pyrubberband==0.3.0
PyScreeze==0.1.28
pyshadow==0.0.5
PySocks==1.7.1
PySoundFile==0.9.0.post1
pyTelegramBotAPI==4.14.0
pytest==7.4.2
python-apt==2.5.3+ubuntu1
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-decouple==3.8
python-dotenv==0.21.0
python-editor==1.0.4
python-telegram-bot==13.15
python3-xlib==0.15
pytweening==1.0.4
pytz==2022.6
pyxattr==0.8.0
pyxdg==0.28
PyYAML==6.0
readchar==4.0.5
referencing==0.30.2
regex==2023.10.3
replicate==0.15.4
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
revChatGPT==6.8.6
rich==13.3.1
rpds-py==0.10.6
rsa==4.9
scraperapi-sdk==0.2.2
SecretStorage==3.3.3
selenium==4.6.0
selenium-wire==5.1.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.3.2.post1
speedtest-cli==2.1.3
starlette==0.27.0
systemd-python==235
telegram==0.0.1
TelegramBotAPI==0.3.2
Telethon==1.32.0
tiktoken==0.5.1
tls-client==0.2.2
toml==0.10.2
tornado==6.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.8.0
tzlocal==5.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
undetected-chromedriver==3.1.7
uritemplate==4.1.1
urllib3==1.26.12
usb-creator==0.3.16
uvicorn==0.24.0.post1
wadllib==1.3.6
waitress==2.1.2
wcwidth==0.2.8
webdriver-manager==3.8.4
websockets==10.4
Werkzeug==3.0.0
wget==3.2
wsproto==1.2.0
xdg==5
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==3.0.3
yarl==1.9.2
zipp==1.0.0
zstandard==0.21.0
